个人业绩能力素质，占比权重改为Bigdecimal，参与考核人员非空判断

+++ src/main/java/com/authine/cloudpivot/web/api/bean/deputyassess/LaunchJudges.java
@@ -3,6 +3,7 @@ package com.authine.cloudpivot.web.api.bean.deputyassess;
 import com.authine.cloudpivot.web.api.bean.User;
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 /**
@@ -14,7 +15,7 @@ import java.util.List;
 public class LaunchJudges {
     private String id;
     private List<User> judges;
-    private int weight;
+    private BigDecimal weight;
     private String roleName;
     private String mutual;
     private String table;
